fix: another cleanup of git.rs
